requirements.txt
pandas
numpy
matplotlib
pillow
scikit-learn
torch
torchvision
tqdm
